+++ toffy/detector_sweep.py
