+++ toffy/file_hash.py
